Modernize AMM5 for Anno 1404 (Classic + History Edition)
- Build: all projects to .NET Framework 4.7.2, clear v4.0 Client profile,
  fix DevelopmentTools build break (legacy AnnoModificationManager classes
  compiled directly instead of a fragile project reference), add .gitignore
  and untrack committed bin/obj/.vs artifacts.
- Version detection: detect game by EXE name/existence instead of exact file
  size (patch-proof); fully wire History Edition / HistoryEdition_Addon into
  the Startup and Settings dialogs; map HE to classic mod compatibility.
- Mod browser: replace the dead bplaced server with a GitHub repository
  (Git Trees API + raw.githubusercontent, SHA cache, derived download URLs,
  TLS 1.2, offline fallback).
- Publishing: DevelopmentTool writes <stem>.xml metadata into a local clone
  of the GitHub repo (user commits/pushes); removed bplaced upload/delete.
- RDA: replace the outdated bundled RDAExplorer with RDAExplorer 1.4.0.0 plus
  a thin compatibility shim (ported RDAManager, public rdaFileEntries,
  RDAReader(string) ctor, RDAFile.SetData, RDAFolder()/1-arg GenerateFrom,
  RDAWriter 3-arg Write + null-tolerant originalReader, RDAFile.Create 2-arg)
  so the manager, DevTool and the bundled GUI build unchanged.
- Tooling: generated ModificationInfo .xml for the existing repo mods
  (Modifications-generated-xml/, also pushed to the Modifications repo).

NOTE: RDA write path is not yet runtime-tested against the game.

+++ a1404mm-code-r23/AnnoModificationManager5/AnnoModificationManager5/ModificationTypes/Modification.cs
@@ -1014,7 +1014,7 @@ namespace AnnoModificationManager5.ModificationTypes
                         {
                             FileInfo fi = new FileInfo(absfile);
 
-                            if (fi.Length == existing.UncompressedSize)
+                            if ((ulong)fi.Length == existing.UncompressedSize)
                             {
                                 try
                                 {
